Introduce very simple authenticity/authorisation
Authenticity token (set in session, sent with each request) makes sure that
message is from the client to whom we sent the token. This defends against
Cross Site Request Forgery.

Authorisiation is currently (mis)managed by taking the user's word for who
they are, and then holding them to that, using the session. Makes it slightly
harder to cheat by changing the handle in the data-handle attribute or by sending
a JS request to the websocket with the wrong game name/an enemy's handle.

Reordered config.ru/app.rb lines so that cookies are set before middleware is
used, which is necessary so that the middleware can use the session.